move `Receipt` protocol and `ReceiptViewModel` protocol into separate files
